Ack messages

+++ app/src/main/java/one/mixin/android/job/BlazeMessageService.kt
@@ -24,10 +24,12 @@ import one.mixin.android.MixinApplication
 import one.mixin.android.R
 import one.mixin.android.api.NetworkException
 import one.mixin.android.api.WebSocketException
+import one.mixin.android.crypto.Base64
 import one.mixin.android.db.FloodMessageDao
 import one.mixin.android.db.JobDao
 import one.mixin.android.db.MixinDatabase
 import one.mixin.android.db.findAckJobsDeferred
+import one.mixin.android.db.findDesktopAckJobsDeferred
 import one.mixin.android.db.findFloodMessageDeferred
 import one.mixin.android.db.findSessionAckJobsDeferred
 import one.mixin.android.di.type.DatabaseCategory
@@ -44,8 +46,12 @@ import one.mixin.android.websocket.BlazeAckMessage
 import one.mixin.android.websocket.BlazeMessage
 import one.mixin.android.websocket.BlazeMessageData
 import one.mixin.android.websocket.ChatWebSocket
+import one.mixin.android.websocket.PlainDataAction
+import one.mixin.android.websocket.TransferPlainAckData
 import one.mixin.android.websocket.createAckListParamBlazeMessage
 import one.mixin.android.websocket.createAckSessionListParamBlazeMessage
+import one.mixin.android.websocket.createParamSessionMessage
+import one.mixin.android.websocket.createPlainJsonParam
 import org.jetbrains.anko.notificationManager
 import java.util.concurrent.Executors
 import javax.inject.Inject
@@ -205,6 +211,9 @@ class BlazeMessageService : Service(), NetworkEventProvider.Listener, ChatWebSoc
         ackJob = GlobalScope.launch(ackThread) {
             ackJobBlock()
             ackSessionJobBlock()
+            Session.getExtensionSessionId()?.let {
+                ackDesktopJobBlock(it)
+            }
         }
     }
 
@@ -234,6 +243,30 @@ class BlazeMessageService : Service(), NetworkEventProvider.Listener, ChatWebSoc
                 list.map { gson.fromJson(it.blazeMessage, BlazeAckMessage::class.java) }.let {
                     try {
                         deliver(createAckSessionListParamBlazeMessage(it)).let {
+                            jobDao.deleteList(list)
+                        }
+                    } catch (e: Exception) {
+                        runAckJob()
+                    } finally {
+                        ackJob = null
+                    }
+                }
+            }
+        }
+    }
+
+    private suspend fun ackDesktopJobBlock(sessionId: String) {
+        jobDao.findDesktopAckJobsDeferred().await()?.let { list ->
+            if (list.isNotEmpty()) {
+                list.map { gson.fromJson(it.blazeMessage, BlazeAckMessage::class.java) }.let {
+                    try {
+                        val plainText = gson.toJson(TransferPlainAckData(
+                            action = PlainDataAction.ACKNOWLEDGE_MESSAGE_RECEIPTS.name,
+                            messages = it
+                        ))
+                        val encoded = Base64.encodeBytes(plainText.toByteArray())
+                        val bm = createParamSessionMessage(createPlainJsonParam(accountId!!, accountId, encoded, sessionId))
+                        deliver(bm).let {
                             jobDao.deleteList(list)
                         }
                     } catch (e: Exception) {
